Remove remaining SFS reference (#56)
* Remove dest_path used for SFS [SKIP GA]

* Remove SFS reference

* Remove SFS reference [SKIP GA]

* Remove SFS reference [SKIP GA]

* Remove SFS reference

* Remove sfs_prefix; update tika version to 2.9.2 [SKIP GA]

* Remove sfs_prefix [SKIP GA]

* Remove debug statement

* Update transfer worker to raise exception when attempting to upload an existing file [SKIP PR]

+++ lib/archival_storage_ingest/workers/transfer_worker.rb
@@ -126,6 +126,8 @@ module TransferWorker
     # source is absolute file path of the asset
     # target is s3_key
     def process_file(source:, target:)
+      raise IngestException, "#{target} already exists in #{_platform}" if s3_manager.exists?(key: target)
+
       s3_manager.upload_file(target, source)
     end
 
@@ -170,6 +172,8 @@ module TransferWorker
     # source is absolute file path of the asset
     # target is s3_key
     def process_file(source:, target:)
+      raise IngestException, "#{target} already exists in #{_platform}" if wasabi_manager.exists?(key: target)
+
       wasabi_manager.upload_file(target, source)
     end
